implementing webinterface

+++ VesselExpress/workflow/scripts/frangi3D.py
@@ -14,8 +14,8 @@ if __name__ == '__main__':
 
     parser = argparse.ArgumentParser(description='Computes Frangi filtering on image files of type .tif')
     parser.add_argument('-i', type=str, help='input tif image file to process')
-    parser.add_argument('-sigma_min', type=int, default=2, help='Frangi sigma_min parameter')
-    parser.add_argument('-sigma_max', type=int, default=5, help='Frangi sigma_max parameter')
+    parser.add_argument('-sigma_min', type=float, default=2.0, help='Frangi sigma_min parameter')
+    parser.add_argument('-sigma_max', type=float, default=5.0, help='Frangi sigma_max parameter')
     parser.add_argument('-sigma_steps', type=int, default=3, help='Frangi sigma_steps parameter')
     parser.add_argument('-alpha', type=float, default=0.5, help='Frangi alpha parameter')
     parser.add_argument('-beta', type=float, default=500.0, help='Frangi beta parameter')
